requirements.txt
# Django==5.0.6
Django==4.2.14
django-environ==0.11.2
pillow==10.2.0
whitenoise==6.7.0
django-allauth==0.63.3

# for Render.com
asgiref==3.8.1
dj-database-url==2.2.0
dj-static==0.0.6
django-model-utils==4.5.1
gunicorn==22.0.0
packaging==24.1
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.1
sqlparse==0.5.1
static3==0.7.0
typing_extensions==4.12.2
tzdata==2024.1